IFS-6696 - Integer updated to BigDecimal for fundingLevel, templates updated, tests updated, migration script added.

+++ ifs-data-layer/ifs-data-service/src/main/java/org/innovateuk/ifs/project/financechecks/service/FinanceCheckServiceImpl.java
@@ -144,7 +144,7 @@ public class FinanceCheckServiceImpl extends AbstractProjectServiceImpl implemen
                                 organisationId,
                                 application.getDurationInMonths(),
                                 projectFinance.getTotal(),
-                                BigDecimal.valueOf(projectFinance.getGrantClaimPercentage()),
+                                projectFinance.getGrantClaimPercentage(),
                                 projectFinance.getTotalFundingSought(),
                                 projectFinance.getTotalOtherFunding(),
                                 projectFinance.getTotalContribution(),
