+++ ifs-data-layer/ifs-data-service/src/test/java/org/innovateuk/ifs/testdata/builders/AcademicCostDataBuilder.java
@@ -92,7 +92,7 @@ public class AcademicCostDataBuilder extends BaseDataBuilder<AcademicCostData, A
         });
     }
 
-    public AcademicCostDataBuilder withGrantClaim(Integer grantClaim) {
+    public AcademicCostDataBuilder withGrantClaim(BigDecimal grantClaim) {
         return updateCostItem(GrantClaimPercentage.class, FinanceRowType.FINANCE, existingCost -> {
             existingCost.setPercentage(grantClaim);
             financeRowCostsService.update(existingCost.getId(), existingCost);
