+++ ifs-resources/src/test/java/org/innovateuk/ifs/finance/builder/ApplicationFinanceResourceBuilderTest.java
@@ -3,6 +3,8 @@ package org.innovateuk.ifs.finance.builder;
 import org.innovateuk.ifs.finance.resource.ApplicationFinanceResource;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 import static org.innovateuk.ifs.finance.builder.ApplicationFinanceResourceBuilder.newApplicationFinanceResource;
 import static org.junit.Assert.assertEquals;
 
@@ -10,7 +12,7 @@ public class ApplicationFinanceResourceBuilderTest {
 
     @Test
     public void testSetGrantClaimPercentage() {
-        ApplicationFinanceResource finance = newApplicationFinanceResource().withGrantClaimPercentage(25).build();
+        ApplicationFinanceResource finance = newApplicationFinanceResource().withGrantClaimPercentage(BigDecimal.valueOf(25)).build();
         assertEquals(25, finance.getGrantClaimPercentage());
     }
 }
